requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
sqlalchemy==2.0.28
plotly==5.19.0
scikit-learn==1.4.1.post1
prophet==1.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openpyxl==3.1.2
pydantic==2.6.3
reportlab==4.1.0
streamlit-authenticator==0.3.1
loguru==0.7.2
